Move dread functions to a separate lib
